Switch to internal mirror and branch tag
Prior to this commit, the Github Action yaml was pulled from this branch
but the base container, and thus the ruby code was pulled from the
latest container which is essentially `master`. This commit switches to
the internal mirror and its copy of the manually pushed tag on docker
hub. e.g. `make push`